pyproject.toml
[project]
name = "uao-neumonia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "img2pdf>=0.4.4",
    "matplotlib>=3.7.3,<4.0.0",
    "numpy>=1.26.0,<2.0.0",
    "opencv-python-headless>=4.9.0.80",
    "pandas>=2.0.3,<3.0.0",
    "pillow>=9.5.0,<11.0.0",
    "plotly>=5.18.0,<6.0.0",
    "pyautogui>=0.9.53",
    "pydicom>=2.4.0",
    "pytest>=7.4.0,<8.0.0",
    "python-xlib>=0.33",
    "scikit-learn>=1.2.2,<2.0.0",
    "scipy>=1.10.1,<2.0.0",
    "seaborn>=0.12.2,<1.0.0",
    "tensorflow==2.16.1",
    "tkcap>=0.0.4",
]
